requirements.txt
fastapi
uvicorn
jinja2
requests
sqlalchemy
psycopg2-binary
python-dotenv